+++ src/index.ts
@@ -50,11 +50,13 @@ export async function run() {
       await restoreCache(primaryKey, [
         igorSetup.runtimeDir,
         igorSetup.bootstrapperDir,
+        igorSetup.userDir
       ]);
     }
 
     await igorSetup.ensureIgorBootStrapperBasedOnOs();
     await igorSetup.getIgorLicense();
+    igorSetup.populateUserDir();
     igorSetup.installModules(targetModulesSplitAsArray);
     core.info(`Installed modules: ${igorSetup.targetModules.join(",")}`);
     core.info(`For runtime: ${targetRuntime}`);